Change Tapper options
- Initial interval: 1 sec -> 0
- Auto reset interval: over 1 sec -> 4 sec

+++ Bipi!/BpmAnimationView.swift
@@ -29,7 +29,7 @@ struct BpmAnimation_Previews: PreviewProvider {
     }
 
     struct PreviewWrapper: View {
-        @State var interval: Double = Double(1000)
+        @State var interval: Double = Double(0)
 
         var body: some View {
             BpmAnimationView(interval: $interval)
